requirements.txt
jupyter
python-chess